Программа разбита на части, исправлено несколько багов
